feat: extract authentication system into @workspace/auth package
