+++ apps/vbcall/package.json
@@ -19,6 +19,7 @@
     "@radix-ui/react-icons": "1.3.2",
     "@t3-oss/env-nextjs": "0.13.8",
     "@tanstack/table-core": "8.21.3",
+    "@workspace/auth": "workspace:*",
     "@workspace/ui": "workspace:*",
     "@workspace/utils": "workspace:*",
     "@workspace/vbcall-db": "workspace:*",
