+++ packages/auth/eslint.config.js
@@ -1,0 +1,4 @@
+import { config } from "@workspace/eslint-config/react-internal";
+
+/** @type {import("eslint").Linter.Config} */
+export default config;
