+++ packages/auth/package.json
@@ -1,0 +1,38 @@
+{
+  "name": "@workspace/auth",
+  "version": "0.0.0",
+  "type": "module",
+  "private": true,
+  "scripts": {
+    "lint": "eslint . --max-warnings 0"
+  },
+  "dependencies": {
+    "@hookform/resolvers": "5.1.1",
+    "@workspace/ui": "workspace:*",
+    "@workspace/utils": "workspace:*",
+    "lucide-react": "0.525.0",
+    "qrcode": "1.5.4",
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
+    "react-hook-form": "7.60.0",
+    "sonner": "2.0.6",
+    "zod": "3.25.76"
+  },
+  "devDependencies": {
+    "@types/qrcode": "1.5.5",
+    "@types/react": "19.1.8",
+    "@types/react-dom": "19.1.6",
+    "@workspace/eslint-config": "workspace:*",
+    "@workspace/typescript-config": "workspace:*"
+  },
+  "exports": {
+    "./components/*": "./src/components/*.tsx",
+    "./components/auth/*": "./src/components/auth/*.tsx",
+    "./hooks/*": "./src/hooks/*.ts",
+    "./hooks/auth/*": "./src/hooks/auth/*.ts",
+    "./lib/*": "./src/lib/*.ts",
+    "./lib/auth/*": "./src/lib/auth/*.ts",
+    "./types/*": "./src/types/*.ts",
+    "./types/auth/*": "./src/types/auth/*.ts"
+  }
+}
