+++ packages/auth/src/components/auth/change-password-form.tsx
@@ -2,30 +2,23 @@
 
 import { Loader2 } from "lucide-react";
 
-import { useResetPasswordForm } from "../../hooks/auth/use-reset-password-form";
-import { Button } from "../button";
-import { ErrorDialog } from "../error-dialog";
-import { Form } from "../form";
-import { FormInputOtp } from "../form/form-input-otp";
-import { FormPasswordInput } from "../form/form-password-input";
+import { Button } from "@workspace/ui/components/button";
+import { ErrorDialog } from "@workspace/ui/components/error-dialog";
+import { Form } from "@workspace/ui/components/form";
+import { FormPasswordInput } from "@workspace/ui/components/form/form-password-input";
 
-type ConfirmResetPasswordFunction = (params: {
-  username: string;
-  confirmationCode: string;
-  newPassword: string;
-}) => Promise<void>;
+import { useChangePasswordForm } from "../../hooks/auth/use-change-password-form";
+import { ConfirmSignInFunction, SignInResult } from "../../types/auth/sign-in";
 
-type props = {
-  setEmailForReset: React.Dispatch<React.SetStateAction<string | null>>;
-  emailForReset: string;
-  confirmResetPasswordFn: ConfirmResetPasswordFunction;
+type props<T extends SignInResult> = {
+  setCurrentState: React.Dispatch<React.SetStateAction<T | null>>;
+  confirmSignInFn: ConfirmSignInFunction<T>;
 };
 
-export function ResetPasswordForm({
-  setEmailForReset,
-  emailForReset,
-  confirmResetPasswordFn,
-}: props) {
+export function ChangePasswordForm<T extends SignInResult = SignInResult>({
+  setCurrentState,
+  confirmSignInFn,
+}: props<T>) {
   const {
     form,
     onSubmit,
@@ -34,11 +27,7 @@ export function ResetPasswordForm({
     closeErrorDialog,
     errorMsg,
     errorTitle,
-  } = useResetPasswordForm({
-    setEmailForReset,
-    emailForReset,
-    confirmResetPasswordFn,
-  });
+  } = useChangePasswordForm({ setCurrentState, confirmSignInFn });
 
   return (
     <Form {...form}>
@@ -51,11 +40,6 @@ export function ResetPasswordForm({
         />
       )}
       <form onSubmit={form.handleSubmit(onSubmit)} className="grid gap-2 py-4">
-        <FormInputOtp
-          control={form.control}
-          name="code"
-          label="Enter the code sent to your email."
-        />
         <FormPasswordInput
           control={form.control}
           name="password"
@@ -74,10 +58,10 @@ export function ResetPasswordForm({
           {isLoading ? (
             <>
               <Loader2 className="mr-2 h-4 w-4 animate-spin" />
-              {"Changing password..."}
+              {"Updating password..."}
             </>
           ) : (
-            "Change password"
+            "Update password"
           )}
         </Button>
       </form>
