+++ packages/auth/src/hooks/auth/use-qr-code.ts
