+++ packages/auth/src/lib/auth/password-validation.ts
