+++ packages/auth/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "@workspace/typescript-config/react-library.json",
+  "include": [
+    "."
+  ],
+  "exclude": [
+    "dist"
+  ]
+}
