+++ packages/ui/package.json
@@ -63,12 +63,8 @@
     "./postcss.config": "./postcss.config.mjs",
     "./tailwind.config": "./tailwind.config.ts",
     "./lib/*": "./src/lib/*.ts",
-    "./lib/auth/*": "./src/lib/auth/*.ts",
     "./components/*": "./src/components/*.tsx",
-    "./components/auth/*": "./src/components/auth/*.tsx",
     "./hooks/*": "./src/hooks/*",
-    "./hooks/auth/*": "./src/hooks/auth/*.ts",
-    "./types/*": "./src/types/*.ts",
-    "./types/auth/*": "./src/types/auth/*.ts"
+    "./types/*": "./src/types/*.ts"
   }
 }
